add tests for theming/custom styles | refactor
