+++ src/test-helpers.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { render } from '@testing-library/react';
 import { ThemeProvider } from 'styled-components';
 import 'jest-styled-components';
-import { generateStyles } from './theme/default';
+import { createStyles } from './DataTable/styles';
 
 export const renderWithTheme = (tree, ...args) =>
-  render(<ThemeProvider theme={generateStyles()}>{tree}</ThemeProvider>, ...args);
+  render(<ThemeProvider theme={createStyles()}>{tree}</ThemeProvider>, ...args);
